Add files via upload
This is the first step of the project, which involves analyzing a heart disease dataset. In this step, I used Excel to clean the data by adding filters to all the column headers. I then examined each column and discovered that some VARCHAR columns contained numbers instead of the expected text values, which I corrected accordingly. Additionally, I noticed that some INTEGER columns had unknown values. To address this, I calculated the average for each affected column and replaced the missing values with the corresponding average. I went ahead to bolden the headings and then added border to beautify the file.